backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6
python-dotenv==1.0.1
python-multipart==0.0.19
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.10.3
pydantic-settings==2.7.0
redis==5.2.1
openai==1.58.1
httpx==0.28.1
python-magic==0.4.27
chardet==5.2.0
pdfplumber==0.11.5
python-docx==1.1.2
openpyxl==3.1.5
beautifulsoup4==4.12.3
pypdf==5.1.0
jieba==0.42.1
langchain-text-splitters==0.3.3
tenacity==9.0.0
